Fix Interjanus channel sync issue
When linking a channel to a remote janus instance's network, the remote server
sends an LSYNC with a new channel. This overwrites the current channel, which
loses users if the channel was already linked across both instances. Fix by
detecting when the keyname of a channel already exists in an LSYNC and replace
with the local copy rather than the constructed copy.