Cache event hooks
This should speed up the parsing of events by not needing to
walk the hook_mod hash every time an action is processed.

As a result, the 'jparse' action is merged into the 'parse' action,
which is now also run on locally-generated actions. Check the
'except' field of the action to differentiate between the cases.