Remove channel locking
Channel locking has caused more problems than it solved.
Replace with an explicit home network for a channel, which
is the unique name for the channel, and whose server is in
charge of the channel merge.

This also allows for a list of "available to link" channels
to be generated/maintained.