Avoid sending double-joins on a channel link
Previously, I was sending double-joins on every link just to be
certain that all servers had the same client list. This isn't
needed, and the cause of the desync (from not sending to remote
janus servers) has been covered. End result: removes the apparent
re-op and re-voice of most people in the channel when a link occurs.