Simplify channel timesync action
Previously, the TIMESYNC action would optionally reset channel
timetamp or wipe modes; this resulted in many special cases.

Since the only time we generate the event is pulling back a TS
on a linked channel, provide for exactly that case, and implement
the others using a MODE plus the timestamp change.

This also fixes some bugs that inspircd modules had with the old code
(timestamp resets always forced a wipe).