Add two-process select loop
This holds all I/O buffers and SSL state in one process, and
does janus processing in another. In the future, the state of the
executing janus process can be frozen and a new secondary process
started with this state. This will allow reclaiming perl memory
leaks or heap fragmentation with no netsplit.